Revert "Forget to commit"
This reverts commit 3f44a7839cd4e386a54f1cff1e0bd1da40a939b0.

Revert "Apply clang-format"

This reverts commit b59bfc0d788b2e8856549df677099f3d0c9f9e18.

+++ Dev/Cpp/Viewer/dll_cs.h
@@ -11,59 +11,52 @@
 #ifndef SWIG_EffekseerNative_WRAP_H_
 #define SWIG_EffekseerNative_WRAP_H_
 
-class SwigDirector_GUIManagerCallback : public efk::GUIManagerCallback, public Swig::Director
-{
+class SwigDirector_GUIManagerCallback : public efk::GUIManagerCallback, public Swig::Director {
 
 public:
-	SwigDirector_GUIManagerCallback();
-	virtual ~SwigDirector_GUIManagerCallback();
-	virtual void Resized(int x, int y);
-	virtual void Droped();
-	virtual void Focused();
-	virtual bool Closing();
-	virtual void Iconify(int f);
-	virtual void DpiChanged(float scale);
-	virtual bool ClickLink(char16_t const* path);
+    SwigDirector_GUIManagerCallback();
+    virtual ~SwigDirector_GUIManagerCallback();
+    virtual void Resized(int x, int y);
+    virtual void Droped();
+    virtual void Focused();
+    virtual bool Closing();
+    virtual void Iconify(int f);
+    virtual void DpiChanged(float scale);
+    virtual bool ClickLink(char16_t const *path);
 
-	typedef void(SWIGSTDCALL* SWIG_Callback0_t)(int, int);
-	typedef void(SWIGSTDCALL* SWIG_Callback1_t)();
-	typedef void(SWIGSTDCALL* SWIG_Callback2_t)();
-	typedef unsigned int(SWIGSTDCALL* SWIG_Callback3_t)();
-	typedef void(SWIGSTDCALL* SWIG_Callback4_t)(int);
-	typedef void(SWIGSTDCALL* SWIG_Callback5_t)(float);
-	typedef unsigned int(SWIGSTDCALL* SWIG_Callback6_t)(void*);
-	void swig_connect_director(SWIG_Callback0_t callbackResized,
-							   SWIG_Callback1_t callbackDroped,
-							   SWIG_Callback2_t callbackFocused,
-							   SWIG_Callback3_t callbackClosing,
-							   SWIG_Callback4_t callbackIconify,
-							   SWIG_Callback5_t callbackDpiChanged,
-							   SWIG_Callback6_t callbackClickLink);
+    typedef void (SWIGSTDCALL* SWIG_Callback0_t)(int, int);
+    typedef void (SWIGSTDCALL* SWIG_Callback1_t)();
+    typedef void (SWIGSTDCALL* SWIG_Callback2_t)();
+    typedef unsigned int (SWIGSTDCALL* SWIG_Callback3_t)();
+    typedef void (SWIGSTDCALL* SWIG_Callback4_t)(int);
+    typedef void (SWIGSTDCALL* SWIG_Callback5_t)(float);
+    typedef unsigned int (SWIGSTDCALL* SWIG_Callback6_t)(void *);
+    void swig_connect_director(SWIG_Callback0_t callbackResized, SWIG_Callback1_t callbackDroped, SWIG_Callback2_t callbackFocused, SWIG_Callback3_t callbackClosing, SWIG_Callback4_t callbackIconify, SWIG_Callback5_t callbackDpiChanged, SWIG_Callback6_t callbackClickLink);
 
 private:
-	SWIG_Callback0_t swig_callbackResized;
-	SWIG_Callback1_t swig_callbackDroped;
-	SWIG_Callback2_t swig_callbackFocused;
-	SWIG_Callback3_t swig_callbackClosing;
-	SWIG_Callback4_t swig_callbackIconify;
-	SWIG_Callback5_t swig_callbackDpiChanged;
-	SWIG_Callback6_t swig_callbackClickLink;
-	void swig_init_callbacks();
+    SWIG_Callback0_t swig_callbackResized;
+    SWIG_Callback1_t swig_callbackDroped;
+    SWIG_Callback2_t swig_callbackFocused;
+    SWIG_Callback3_t swig_callbackClosing;
+    SWIG_Callback4_t swig_callbackIconify;
+    SWIG_Callback5_t swig_callbackDpiChanged;
+    SWIG_Callback6_t swig_callbackClickLink;
+    void swig_init_callbacks();
 };
 
-class SwigDirector_IOCallback : public Effekseer::IOCallback, public Swig::Director
-{
+class SwigDirector_IOCallback : public Effekseer::IOCallback, public Swig::Director {
 
 public:
-	SwigDirector_IOCallback();
-	virtual void OnFileChanged(Effekseer::StaticFileType fileType, char16_t const* path);
+    SwigDirector_IOCallback();
+    virtual void OnFileChanged(Effekseer::StaticFileType fileType, char16_t const *path);
 
-	typedef void(SWIGSTDCALL* SWIG_Callback0_t)(int, void*);
-	void swig_connect_director(SWIG_Callback0_t callbackOnFileChanged);
+    typedef void (SWIGSTDCALL* SWIG_Callback0_t)(int, void *);
+    void swig_connect_director(SWIG_Callback0_t callbackOnFileChanged);
 
 private:
-	SWIG_Callback0_t swig_callbackOnFileChanged;
-	void swig_init_callbacks();
+    SWIG_Callback0_t swig_callbackOnFileChanged;
+    void swig_init_callbacks();
 };
+
 
 #endif
